First draft complete

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bpuschel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/26 14:09:26 by bpuschel          #+#    #+#             */
-/*   Updated: 2017/05/29 19:43:28 by bpuschel         ###   ########.fr       */
+/*   Updated: 2017/05/31 21:24:15 by bpuschel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,6 +47,6 @@ int				ft_printf(char *fmt, ...);
 char			*ft_itoa_base(intmax_t value, int base, int precision);
 int				num_handler(char c, va_list *args, int **flags, int **wpl);
 int				chr_handler(char c, va_list *args, int **flags, int **wpl);
-char			*ft_strfill(char *src, const char *fill, int len, int dir);
+char			*ft_strfill(char *src, const char *fill, size_t len, int dir);
 
 #endif
